refactor: train setup/sounds/state split
