+++ CMakeLists.txt
@@ -98,6 +98,7 @@ set(SOURCES
 "world/Traction.cpp"
 "world/TractionPower.cpp"
 "vehicle/train/Train.cpp"
+"vehicle/train/TrainCabSetup.cpp"
 "vehicle/train/TrainCommandsAuxiliary.cpp"
 "vehicle/train/TrainCommandsBrakes.cpp"
 "vehicle/train/TrainCommandsControls.cpp"
@@ -106,6 +107,8 @@ set(SOURCES
 "vehicle/train/TrainCommandsLights.cpp"
 "vehicle/train/TrainCommandsMisc.cpp"
 "vehicle/train/TrainCommandsRadio.cpp"
+"vehicle/train/TrainSounds.cpp"
+"vehicle/train/TrainState.cpp"
 "world/TrkFoll.cpp"
 "vehicle/AirCoupler.cpp"
 "model/AnimModel.cpp"
